+++ Sources/picoLib/Instructions/STR.swift
@@ -36,7 +36,7 @@ public struct STR: Instruction {
         throw ParserError.unexpectedError
     }
 
-    public func encode(symbols: [String: UInt16]) throws -> [UInt16] {
+    public func encode(symbols: [String: Int]) throws -> [UInt16] {
         switch kind {
         case let .STR_Immediate_T1(r1, r2, imm):
             return Thumb.STR_Immediate_T1(t: r1.number, n: r2.number, imm5: imm).encode()
